requirements.txt
matplotlib==3.7.1
matplotlib==3.7.0
numpy==1.23.5
numpy==1.22.3
pandas==2.0.0
pandas==1.5.3
scikit_learn==1.2.1
tensorflow_macos==2.12.0
